A few changes

The result is going from
```
satisfia (b28010a) 📦🤷 🐍 v3.10.13 via ⊢💩 (shell) on ☁️  m.redaelli@gmail.com
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
  Time (mean ± σ):     38.970 s ±  0.577 s    [User: 38.611 s, System: 0.551 s]
  Range (min … max):   38.037 s … 39.630 s    10 runs
```
to
```
satisfia on 🌱 perf 📦🤷 🐍 v3.10.13 via ⊢💩 (shell) on ☁️  m.redaelli@gmail.com
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
  Time (mean ± σ):     30.802 s ±  1.359 s    [User: 30.457 s, System: 0.587 s]
  Range (min … max):   29.625 s … 33.938 s    10 runs
```

last commit goes down to 24 secs