@mredaelli 2024-03-30
Caching also `propagateAspiration` brings it down to 29s, and the memory usage for stays below 450MB.