@mredaelli 2024-03-30
Next things to consider would be:

- `_extract_state_attributes` takes a lot of time. If `state` instead were already a tuple, we'd save all that computation plus the time to call the function. Would that be messy?
- `transition_distribution`