@mredaelli 2024-04-06
I pushed a cleanup of the utilities, with the interval version and the "scalar but maybe second argument is an interval" version:

```
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
  Time (mean ± σ):     26.851 s ±  0.742 s    [User: 26.346 s, System: 0.600 s]
  Range (min … max):   26.273 s … 28.715 s    10 runs
```

so we are saving about 30%.

I actually specialized the utilities a bit more, adding a scalar version (e.g. `clip`), a version where only the central argument is a tuple (`clip2`) and a generic one for intervals (`clipInt`), the original one, which however doesn't seem to be used.

That saves a bit more, and I like it better because it's more... "typed" :), but I understand if it gets in the way:
```
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
 ⠦ Current estimate: 26.614 s     ████████████████████████████░░░░░░░░░░░░░░░░░░░░ ETA 00:01:46  Time (mean ± σ):     26.436 s ±  0.631 s    [User: 25.798 s, System: 0.607 s]
  Range (min … max):   25.509 s … 27.904 s    10 runs
```