@mredaelli 2024-04-07
Latest commit shaves another couple of seconds by excluding from the loss computation stuff with 0 loss parameter directly in `__init__`. 

Losses for time and `lossCoeff4OtherLoss` still need adaptation though.

I also replaces all tabs with spaces because I'm going crazy with whitespace errors. I hope that's OK