@mensch72 2024-04-09
parameters cannot be changed during the lifetime of the agent at the moment.

Am 9. April 2024 20:17:52 MESZ schrieb Massimo Redaelli ***@***.***>:
>@mredaelli commented on this pull request.
>
>
>
>> +        return res
>+
>+    @cache
>+    def randomTieBreaker(self, state, action):
>+        return random.random()
>+
>+    # now we can combine all of the above quantities to a combined (safety) loss function:
>+
>+    # state, action, aleph4state, aleph4action, estActionProbability
>+    @lru_cache(maxsize=None)
>+    def combinedLoss(self, s, a, al4s, al4a, p): # recursive
>+
>+        if self.debug:
>+            print(pad(s),"| | combinedLoss, state",prettyState(s),"action",a,"aleph4state",al4s,"aleph4action",al4a,"estActionProbability",p,"...")
>+
>+        res = sum(k * v(s, a, al4s, al4a, p) for k, v in self.losses.items())
>
>If I'm not mistaken, that's done in `__init__`, here:
>
>https://github.com/pik-gane/satisfia/pull/7/files#diff-2a77c67961414a5d12c1a7db8c3b4536e945a06643c4f16f68f3509b002378eeR190
>
>so we do it only once rather than at every evaluation.
>
>Unless the parameters can be changed?
>
>-- 
>Reply to this email directly or view it on GitHub:
>https://github.com/pik-gane/satisfia/pull/7#discussion_r1558109685
>You are receiving this because your review was requested.
>
>Message ID: ***@***.***>