@mredaelli 2024-04-13
rebased and removed some numpy.

Down to:
```
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
  Time (mean ± σ):     22.559 s ±  0.579 s    [User: 21.415 s, System: 1.334 s]
  Range (min … max):   21.661 s … 23.575 s    10 runs
```