@mredaelli 2024-04-13
Added a commit to remove `prev_loc` from state.

```
❯ hyperfine "python scripts/perf_eval.py"
Benchmark 1: python scripts/perf_eval.py
  Time (mean ± σ):      5.990 s ±  0.359 s    [User: 5.750 s, System: 0.521 s]
  Range (min … max):    5.670 s …  6.830 s    10 runs
```